DRY: extract shared FormField parser (was duplicated in form.ts + subform.ts)
Audit called this out: parseFormField in form.ts and parseField in
subform.ts were near-identical 60-line copies of the same logic. The
form.ts version even had a comment "Same field parsing as subform.ts
(kept inline to avoid circular imports)".

Extracted to a new field-parser.ts module. Both form.ts and subform.ts
now import parseFormField from this shared module. No circular
dependency because field-parser.ts imports only from tokenize (which
has no upward deps) and from types/Form (type-only).

Behavior preserved:
- Same field defaults
- Same keyword handlers (label, definition, unit, required,
  measurement_method, calculation, calculation_bindings, derivation,
  evaluation, values, default, min_items, max_items, items, fields,
  reference)
- Same forward-compat skip for unknown keywords
- stripWrapping for bare-ID fields (calculation)
- removePackage for quoted-string fields (label, definition, etc.)

Lines removed: ~120 (60 from each file).
Lines added: ~70 (the new shared module).

Net: ~50 fewer lines, single source of truth for field parsing.

Tests: 77 pass (unchanged). R 60 model still parses cleanly: 41 forms,
1 subform, all field structures intact.
